rename `ciphers.rs` to `cipher.rs`
