+++ src/crypto/decrypt.rs
@@ -1,5 +1,5 @@
 use crate::crypto::key::argon2_hash;
-use crate::crypto::primitives::ciphers::Ciphers;
+use crate::crypto::primitives::cipher::Ciphers;
 use crate::global::header::Header;
 use crate::global::secret::Secret;
 use aead::Payload;
