+++ src/crypto/primitives/cipher.rs
